feat(contentlayout): update contentLayout export table/header
